Release 1.0.1
The most recent changes to the code are intended to fix what I see as an issue in existing Bowser backend logic. Uploads are meant to be idempotent, but the existing logic did not take in to account file names that are temporally unique like UUID or timestamp based names. This is seen as a fix for that, hence the patch version bump.